requirements.txt
streamlit==1.37.1
python-docx==1.1.2
reportlab==4.2.2
